Added the official CPU constitution
The official version of the constitution is currently out of date,
however, in order to be incompliance with the versioning and
transparency clauses, it is necessary to commit the old version of the
constitution first. Future modifications to this document must go
through the official amendment and ratification process specified
herein.

On a meta note, the git commit of each modification to this document
should contain copious details for posterity. In this regard, this
commit should serve as an example.

Additionally, this git commit's author timestamp *has* been tampered
with to reflect the exact date and time the constitution was last
edited. Since the anti-tampering provisions of the constitution are not
yet official, this should be constitutionally valid. Additionally, the
committer's timestamp has not been tampered with.