Added a transparency clause to the Constitution
Historically, CPU has been quite opaque. In fact, the constitution has
typically not been made available to club members. This clause seeks to
change how the club is run by ensuring that everybody is able to easily
read and suggest changes to the constitution. Additionally, by forcing
all changes to be properly versioned, it should be easier for club i
members to see the history of the constitution.